chore: rebase & remove conflicts
