+++ core/color/color.go
