+++ core/factory/factory.go
@@ -2,6 +2,7 @@ package factory
 
 import (
 	"github.com/aerogear/charmil/core/config"
+	"github.com/aerogear/charmil/core/iostreams"
 	"github.com/aerogear/charmil/core/localize"
 	"github.com/aerogear/charmil/core/logging"
 )
@@ -9,6 +10,8 @@ import (
 // Factory is an abstract type which provides
 // the access of charmil packages to the commands
 type Factory struct {
+	// Type which defines the streams for the CLI
+	IOStreams *iostreams.IOStreams
 	// Logger provides functions for unified logging
 	Logger logging.Logger
 	// Localizer localizes the text in commands
