+++ examples/plugins/echo/echo.go
@@ -2,7 +2,6 @@ package echo
 
 import (
 	"embed"
-	"log"
 
 	"github.com/aerogear/charmil/core/factory"
 	"github.com/aerogear/charmil/core/localize"
@@ -23,17 +22,8 @@ type Options struct {
 	Localize localize.Localizer
 }
 
-// Initializes a zero-valued struct and stores its address
-var cfg = &config{}
-
 // Date Command
-func EchoCommand(f *factory.Factory) (*cobra.Command, error) {
-
-	// Stores the config for localizer
-	cfg.LocConfig = localize.Config{
-		Language: language.English,
-		Path:     "examples/plugins/echo/locales/en/en.yaml",
-		Format:   "yaml"}
+func EchoCommand() (*cobra.Command, error) {
 
 	// Initialize localizer providing the language, locals and format of locals file
 	loc, err := localize.New(&localize.Config{
@@ -46,19 +36,20 @@ func EchoCommand(f *factory.Factory) (*cobra.Command, error) {
 		return nil, err
 	}
 
-	opts := &factory.Factory{
-		Logger:     f.Logger,
-		Localizer:  loc,
-		CfgHandler: f.CfgHandler,
+	// Create new/default instance of factory
+	newFactory := factory.Default(loc)
+	opts := &Options{
+		Logger:   newFactory.Logger,
+		Localize: newFactory.Localizer,
 	}
 
 	// creating new echo command
 	// using localizer to access default text by ID provided in locals
 	cmd := &cobra.Command{
-		Use:     opts.Localizer.LocalizeByID("echo.cmd.use"),
-		Short:   opts.Localizer.LocalizeByID("echo.cmd.short"),
-		Long:    opts.Localizer.LocalizeByID("echo.cmd.long"),
-		Example: opts.Localizer.LocalizeByID("echo.cmd.example"),
+		Use:     opts.Localize.LocalizeByID("echo.cmd.use"),
+		Short:   opts.Localize.LocalizeByID("echo.cmd.short"),
+		Long:    opts.Localize.LocalizeByID("echo.cmd.long"),
+		Example: opts.Localize.LocalizeByID("echo.cmd.example"),
 		Args:    cobra.ExactArgs(1),
 		RunE: func(cmd *cobra.Command, args []string) error {
 
@@ -67,12 +58,6 @@ func EchoCommand(f *factory.Factory) (*cobra.Command, error) {
 
 			return nil
 		},
-	}
-
-	// Merges the current plugin config into the host CLI config
-	err = c.MergePluginCfg("echo", opts.CfgHandler, cfg)
-	if err != nil {
-		log.Fatal(err)
 	}
 
 	return cmd, nil
