+++ go.mod
@@ -4,10 +4,13 @@ go 1.16
 
 require (
 	github.com/BurntSushi/toml v0.3.1
+	github.com/fatih/color v1.12.0
 	github.com/imdario/mergo v0.3.12
+	github.com/mattn/go-isatty v0.0.12
 	github.com/nicksnyder/go-i18n/v2 v2.1.2
 	github.com/pelletier/go-toml v1.2.0
 	github.com/spf13/cobra v1.1.3
+	github.com/spf13/pflag v1.0.5
 	github.com/tidwall/sjson v1.1.7
 	golang.org/x/text v0.3.3
 	gopkg.in/yaml.v2 v2.4.0
